feat: stores clips in MongoDB
When a clip is "got", we try to get the clip from the database.
If the clip exists, then perfect, if not, the clip is retreived from
the Twitch API and stored in the DB.
Each clip contains all the information from Twitch, which will be
useful when adding data to stitching (like displaying which user
curated the clip).

https://github.com/Sadzeih/stitchclips/issues/21